Added bfragweights.root, PDF gen tools 	modified:   interface/CharmTree.h 	new file:   interface/GeneratorTools.h 	modified:   src/CharmTree.cc 	new file:   src/GeneratorTools.cc 	modified:   src/TopKalman.cc

+++ TopAnalysis/interface/GeneratorTools.h
@@ -14,7 +14,7 @@
 #include "TF1.h"
 
 //available theory systs
-typedef std::pair<TString, int> WeightSysts_t;
+typedef std::pair<TString, float> WeightSysts_t;
 std::vector< WeightSysts_t > getWeightSysts(TFile *, TString sample="TTJets2016");
 std::vector< WeightSysts_t > getPartonShowerWeightSysts(TFile *f);
